:zap: Use template for StepEvent and some functions
